Şirket sayfası GRİ EKRAN: `adet` metin geliyordu, `as num?` ile okunuyordu
KULLANICI BİLDİRDİ: "Dizi profilinden yapım firmalarını açınca gri bir ekran
çıkıyor, Android cihazımda da webde de aynı."

KÖK NEDEN
`/incelemeler/:tur/:id` ucu `adet`i METİN gönderiyordu ("0"): SQL `count(*)`
bigint döner ve node-postgres bigint'i dizgeye çevirir. `sirket.dart` bunu
`as num?` ile okuyunca `type 'String' is not a subtype of type 'num?'`
fırlıyor ve TÜM sayfa gri ekrana düşüyordu. `ortalama` da `numeric` olduğu
için aynı kapıya çıkıyordu — firma puanlandığı anda ikinci kez çökerdi.

NEDEN TESTLER YAKALAMADI — ASIL HATA BENDE
`sirket_puan_raf_test.dart`in sahte yanıtında `'adet': 0` yazıyordu, yani bir
SAYI. Testi yazarken sunucunun GERÇEK yanıtını kopyalamak yerine şekli
UYDURMUŞTUM; test kendi uydurmamı doğruladı. Üstelik sayfayı tarayıcıda hiç
açmamıştım — `curl` ile 200 almak SPA kabuğunu doğrular, Flutter'ın o kabuğun
içinde çizip çizmediğini DEĞİL. Hatayı kullanıcı bulmak zorunda kaldı.

ÜÇ KATMANDA DÜZELTİLDİ
* İstemci: `puanSayisi()`. Bu yardımcı projede TAM DA bu tuzak için zaten
  vardı ("sunucu kimi uçta sayı, kimi uçta metin gönderiyor") ve kardeş
  ekranlar (kisi.dart, detay.dart) onu kullanıyordu — kopyalarken atlanmış.
* Sunucu: `count(*)::int AS adet`. Aynı ucun kardeş dağılım sorgusu zaten
  `::int` kullanıyordu; bu satır o disiplinin dışında kalmıştı. `ortalama`
  bilerek `numeric` kalıyor (1 ondalık kesinliği float'a çevrilince kaybolur;
  tüm okuyucular `puanSayisi` ile iki biçimi de kabul ediyor).
* Test: sahte yanıt GERÇEK şekle çevrildi ve METİN bırakıldı — istemci her
  iki biçimi de kaldırabilmeli. Düzeltme geri alınınca 4 testin DÖRDÜ de
  kırmızıya döndü, yani test artık gerçekten bekçilik ediyor.

KANIT
Canlı tarayıcıda sayfa açıldı: HBO başlığı, Puanla düğmesi, tepki satırı,
"Devam eden yapımlar (19)", "Diziler (20)" — konsolda hata yok.
1755 Flutter + 1271 backend testi yeşil. Uç artık `"adet":0` değil `"adet":0`
(sayı) dönüyor.

+++ backend/server.js
@@ -5840,7 +5840,15 @@ app.get('/incelemeler/:tur/:tmdbId', girisIsteğeBagli, sarici(async (req, res) 
       [...olcut, req.kullanici?.id || 0],
     ),
     havuz.query(
-      `SELECT round(avg(puan)::numeric, 1) AS ortalama, count(*) AS adet
+      // `::int` ŞART (19 Ağu 2026): `count(*)` bigint döner ve node-pg
+      // bigint'i METİN olarak gönderir — yani `adet` JSON'da "0" idi, 0
+      // değil. Şirket sayfası bunu sayıya çevirmeye çalışınca TÜM SAYFA gri
+      // ekrana düştü. Hemen aşağıdaki dağılım sorgusu zaten `::int`
+      // kullanıyordu; bu satır o disiplinin dışında kalmıştı.
+      // `ortalama` BİLEREK numeric kalıyor: 1 ondalık basamağın kesinliği
+      // float'a çevrilince kaybolur, ve tüm okuyucular zaten `puanSayisi()`
+      // ile hem metni hem sayıyı kabul ediyor.
+      `SELECT round(avg(puan)::numeric, 1) AS ortalama, count(*)::int AS adet
        FROM puanlar WHERE tur=$1 AND tmdb_id=$2 AND sezon IS NULL`,
       olcut,
     ),
